+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 v0.2.0 (in development)
 -----------------------
 - Increased MSRV to 1.70
+- **Breaking**: All error types have been combined into a single
+  `InPlaceError`, and all error kinds have been combined into a single
+  `InPlaceErrorKind`
 
 v0.1.0 (2023-05-17)
 -------------------